Add wildcard and ingress DNS records (#11)
* Refactor reconciler

* replace cloud dns client with registrars
* each registrar registers a separate dns record (or the zone itself)
* registrars are called sequentially in the reconciler
* registrars are deleted in reverse order (since the first registrar
  should always be the zone registrar)

* Add Service client

* Add ingress registrar

Also do not use the go resolver in acceptance tests. This causes lookups to
just fail after a few tests.

* Add wildcard registrar

* Use 2 nodes for acceptance tests

Using more nodes makes the tests flaky. Likely because the dns requests
get throttled after a while

* Don't error if zone ns record already exists

* Bump depdendencies

Explicitly require transitive dependencies with CVEs to make nancy happy
Bump controller-gen to v0.9.0

* Add make target to lint imports

Also lint imports

* Fix ingress registrar not returning error

* Add missing script

* Add more granular logging

* Use the logger in the context instead of passing one to the reconciler
  constructor. This unfortunately makes checking that the context is
  passed not possible in the unit tests (or at least very hard) since
  the context is mutated.
* Add start/done logs to each registrar
* Add more error logs
* Add more info logs when reconciliation is skipped

* Use RFC3339 time format. Remove gcpclsuter value from logger

The value is already provided by controller-runtime

* Wait for cluster deletion after acceptance tests in ci

Sometimes dns records leak after ci runs. Presumably this is because the
tests only trigger the deletion of the cluster CRs when cleaning up, but
don't actually wait for them to be deleted and the Cloud DNS resources
to be cleaned up.

* Fix typo