Improve acceptance tests performance
This both makes the test run faster and reduces the chances of a flake
happening. This is because DNS queries will get throttled and
progressively become slower, until something fails. By reducing the
amount of clusters created to just two, this should also reduce the
amount of Eventuallies that need to poll until a specific domain becomes
available